int32_t -> int64_t

+++ JJLISO8601DateFormatter/Classes/JJLInternal.h
@@ -7,5 +7,5 @@
 static const int32_t kJJLMaxDateLength = 50; // Extra to be safe
 
 void JJLFillBufferForDate(char *buffer, double timeInSeconds, CFISO8601DateFormatOptions options, timezone_t timeZone, double fallbackOffset);
-double JJLTimeIntervalForString(const char *string, int32_t length, CFISO8601DateFormatOptions options, timezone_t timeZone, bool *errorOccurred);
+double JJLTimeIntervalForString(const char *string, int64_t length, CFISO8601DateFormatOptions options, timezone_t timeZone, bool *errorOccurred);
 void JJLPerformInitialSetup(void);
